requirements-test.txt
coverage
coveralls
mock>=1.0.1
nose>=1.3.0
django-nose>=1.2
django-sekizai
https://github.com/divio/django-cms/archive/develop.zip

# Additional test requirements go here